+++ src/inspectors/hues/isPinkHue.ts
@@ -33,7 +33,7 @@
 //
 
 import type { AnyCssColor } from "../../CssColor/AnyCssColor.type";
-import { hues } from "./hues";
+import { isHue } from "../isHue";
 
 /**
  * isPinkHue() returns true if the given `input` color is a pink hue.
@@ -43,10 +43,12 @@ import { hues } from "./hues";
  * @returns
  * - `true` if `input` is a pink hue
  * - `false` otherwise
+ *
+ * @deprecated Use `isHue(input, "pink")` instead.
  */
 export function isPinkHue(
     input: AnyCssColor
 ): boolean
 {
-    return hues(input).some((hue) => hue === "pink");
+    return isHue(input, "pink");
 }